catalog/globalcatalog/tasks/ConvertFileFormat/requirements.txt
PyYAML
toml
xmltodict
openpyxl
numpy
pandas
